Speculative fix for cros_daisy compilation failure
include/core/SkDeviceProperties.h:82:51: error: narrowing conversion of '(((int)orientation) | ((int)layout))' from 'int' to 'uint8_t

unreviewed



git-svn-id: http://skia.googlecode.com/svn/trunk@7329 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkDeviceProperties.h
@@ -79,12 +79,12 @@ struct SkDeviceProperties {
         static Geometry MakeDefault() {
             Orientation orientation = fromOldOrientation(SkFontHost::GetSubpixelOrientation()); //kHorizontal_Orientation
             Layout layout = fromOldLayout(SkFontHost::GetSubpixelOrder()); //kRGB_Layout
-            Geometry ret = { orientation | layout };
+            Geometry ret = { uint8_t(orientation | layout) };
             return ret;
         }
 
         static Geometry Make(Orientation orientation, Layout layout) {
-            Geometry ret = { orientation | layout };
+            Geometry ret = { uint8_t(orientation | layout) };
             return ret;
         }
